+++ apps/api/src/app/auth/mobile/login/[provider]/callback.ts
@@ -1,8 +1,8 @@
 import { AuthUseCase } from "@/application/use-cases/auth";
 import { OAuthUseCase } from "@/application/use-cases/oauth";
 import { OAuthAccountUseCase } from "@/application/use-cases/oauth-account";
-import { providerSchema, selectOAuthProviderGateway } from "@/interface-adapter/gateway/oauth-provider";
-import { LuciaAdapter } from "@/interface-adapter/lucia";
+import { LuciaAdapter } from "@/infrastructure/lucia";
+import { oAuthProviderSchema, selectOAuthProviderService } from "@/infrastructure/oauth-provider";
 import { OAuthAccountRepository } from "@/interface-adapter/repositories/oauth-account";
 import { ElysiaWithEnv } from "@/modules/elysia-with-env";
 import { BadRequestException } from "@/modules/error/exceptions";
@@ -25,7 +25,7 @@ const ProviderCallback = new ElysiaWithEnv().get(
 		const providerGatewayRedirectUri = new URL(`auth/mobile/login/${provider}/callback`, apiBaseUri);
 
 		const oAuthUseCase = new OAuthUseCase(
-			selectOAuthProviderGateway({
+			selectOAuthProviderService({
 				provider,
 				env,
 				redirectUri: providerGatewayRedirectUri.toString(),
@@ -109,7 +109,7 @@ const ProviderCallback = new ElysiaWithEnv().get(
 			{ additionalProperties: true },
 		),
 		params: t.Object({
-			provider: providerSchema,
+			provider: oAuthProviderSchema,
 		}),
 		cookie: t.Cookie({
 			[OAUTH_STATE_COOKIE_NAME]: t.String(),
