+++ apps/api/src/app/auth/mobile/logout.ts
@@ -1,5 +1,5 @@
 import { AuthUseCase } from "@/application/use-cases/auth";
-import { LuciaAdapter } from "@/interface-adapter/lucia";
+import { LuciaAdapter } from "@/infrastructure/lucia";
 import { ElysiaWithEnv } from "@/modules/elysia-with-env";
 import { BadRequestException } from "@/modules/error/exceptions";
 import { t } from "elysia";
