+++ apps/api/src/app/auth/web/login/[provider]/index.ts
@@ -1,5 +1,5 @@
 import { OAuthUseCase } from "@/application/use-cases/oauth";
-import { providerSchema, selectOAuthProviderGateway } from "@/interface-adapter/gateway/oauth-provider";
+import { oAuthProviderSchema, selectOAuthProviderService } from "@/infrastructure/oauth-provider";
 import { ElysiaWithEnv } from "@/modules/elysia-with-env";
 import { BadRequestException } from "@/modules/error/exceptions";
 import {
@@ -24,7 +24,7 @@ const Provider = new ElysiaWithEnv({ prefix: "/:provider" })
 			const providerGatewayRedirectUri = new URL(`auth/web/login/${provider}/callback`, apiBaseUri);
 
 			const oAuthUseCase = new OAuthUseCase(
-				selectOAuthProviderGateway({
+				selectOAuthProviderService({
 					provider,
 					env,
 					redirectUri: providerGatewayRedirectUri.toString(),
@@ -81,7 +81,7 @@ const Provider = new ElysiaWithEnv({ prefix: "/:provider" })
 				"redirect-uri": t.Optional(t.String()),
 			}),
 			params: t.Object({
-				provider: providerSchema,
+				provider: oAuthProviderSchema,
 			}),
 		},
 	)
