+++ apps/api/src/app/auth/web/signup/index.ts
@@ -1,6 +1,6 @@
 import { AuthUseCase } from "@/application/use-cases/auth";
 import { UserUseCase } from "@/application/use-cases/user";
-import { LuciaAdapter } from "@/interface-adapter/lucia";
+import { LuciaAdapter } from "@/infrastructure/lucia";
 import { UserRepository } from "@/interface-adapter/repositories/user";
 import { ElysiaWithEnv } from "@/modules/elysia-with-env";
 import { SESSION_COOKIE_NAME } from "@mona-ca/core/const";
