+++ apps/api/src/infrastructure/lucia/interface/lucia.adapter.interface.ts
