+++ apps/api/src/infrastructure/oauth-provider/index.ts
@@ -1,0 +1,8 @@
+export type {
+	IOAuthProviderService as IOAuthProviderGateway,
+	AccountInfo,
+} from "./interface/oauth-provider.service.interface";
+export { DiscordOAuthService } from "./providers/discord.service";
+export type { OAuthProvider } from "./types/provider.type";
+export { oAuthProviderSchema } from "./schema/provider.schema";
+export { selectOAuthProviderService } from "./utils/select-oauth-provider-service";
