+++ apps/api/src/infrastructure/oauth-provider/interface/oauth-provider.service.interface.ts
@@ -8,7 +8,7 @@ export type AccountInfo = {
 	emailVerified: boolean;
 };
 
-export interface IOAuthProviderGateway {
+export interface IOAuthProviderService {
 	genAuthUrl(state: string, codeVerifier: string): URL;
 	getTokens(code: string, codeVerifier: string): Promise<OAuth2Tokens>;
 	getAccountInfo(accessToken: string): Promise<AccountInfo | null>;
