+++ apps/api/src/infrastructure/oauth-provider/schema/provider.schema.ts
