+++ apps/api/src/infrastructure/oauth-provider/types/provider.type.ts
